Large Overhaul
Currently working to get the DB values to appear on screen. Broke the scraping feature currently in the works for that. Set up routes and struts for the finanical records to start calculating DCF and WAAC.

Also started work on the login page for sign up and sign in. Going to try and create a native authentication service.